+++ packages/core/src/index.ts
@@ -63,4 +63,6 @@ export { fetchWithTimeout, HttpError } from './utils/http'
 export type { FetchWithTimeoutOptions } from './utils/http'
 export { callVisionOcr, DEFAULT_VISION_MODEL, DEFAULT_VISION_CONCURRENCY, DEFAULT_VISION_PROMPT, DEFAULT_VISION_TIMEOUT_MS } from './utils/vision-ocr'
 export type { VisionOcrOptions, VisionOcrResult, VisionOcrFailure } from './utils/vision-ocr'
+export { loadTokensCache, saveTokensCache, TOKENS_FILE } from './utils/chunk-cache'
+export type { PersistedTokens } from './utils/chunk-cache'